+++ examples/dev/demo/lila.config.js
@@ -117,9 +117,9 @@ module.exports = {
     }
   },
   esLint: !0,
-  esLintFix: !0,
+  // esLintFix: !0,
   styleLint: !0,
-  styleLintFix: !0,
+  // styleLintFix: !0,
   prettier: !0,
   prettierOptions: {
     // 'print-width': 40