+++ examples/revert/index-2.js
@@ -1,4 +1,0 @@
-
-require('../../util/change_cwd')(__dirname + '/demo');
-
-require('../../util/exec')('lila revert -i 2');