+++ examples/revert/not-exist.js
@@ -1,4 +1,0 @@
-
-require('../../util/change_cwd')(__dirname);
-
-require('../../util/exec')('lila revert');