+++ tasks/main/arc.js
@@ -1,8 +1,0 @@
-/**
- * Register `arc` task.
- *
- * @param gulp
- */
-module.exports = gulp => {
-  gulp.task('arc', gulp.series('pre_archive'));
-};