+++ test/clean/demo/bak/1.721ea8e8a5ae693fd7ed70b501c7d28c.js
@@ -1,1 +1,0 @@
-console.log('1.721ea8e8a5ae693fd7ed70b501c7d28c');