+++ test/clean/demo/bak/2.e2025f09faac9dd460cbac6913cfbda6.js
@@ -1,1 +1,0 @@
-console.log('2.e2025f09faac9dd460cbac6913cfbda6');